Update: move NMSUtils to nms package
