+++ common/src/main/java/cc/zoyn/core/util/NameTagUtils.java
@@ -1,5 +1,6 @@
 package cc.zoyn.core.util;
 
+import cc.zoyn.core.util.nms.NMSUtils;
 import com.mojang.authlib.GameProfile;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
