+++ common/src/main/java/cc/zoyn/core/util/nms/NMSUtils.java
@@ -1,4 +1,4 @@
-package cc.zoyn.core.util;
+package cc.zoyn.core.util.nms;
 
 import cc.zoyn.core.util.reflect.ReflectionUtils;
 import org.apache.commons.lang3.Validate;
